+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Added ability to specify where the circuit's AST files are stored
     - Added ability to specify where to put generated artifacts and types
 - Implemented Zkit wrapper generation for given circuits
+- Used ejs instead of ts factory to render the wrapper class
 
 ## [v0.1.1]